model/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html

absl-py==0.14.1
asgiref==3.4.1
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
dj-database-url==0.5.0
Django==3.2.8
djangorestframework==3.12.4
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
mediapipe==0.8.8
mysqlclient==2.0.3
numpy==1.21.2
opencv-contrib-python==4.5.3.56
opencv-python-headless==4.5.3.56
packaging==21.0
Pillow==8.4.0
protobuf==3.18.1
psutil==5.8.0
psycopg2==2.9.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
torch==1.9.1+cpu
torchvision==0.10.1+cpu
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
